# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a specialized server for managing trading strategies and margins.",
        "title": "Trades Management API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/nse/allindices": {
            "get": {
                "description": "Fetches latest performance data for all indices using the warmup strategy.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stocks"
                ],
                "summary": "Get All NSE Indices",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.AllIndicesResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/nse/heatmap": {
            "get": {
                "description": "Fetches latest sectoral data. Uses warmup time-cache strategy to avoid NSE rate limits.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stocks"
                ],
                "summary": "Get NSE Sectoral Heatmap",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.SectorData"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/nse/history": {
            "get": {
                "description": "Fetches stock history for a specific symbol. Utilizes a 1-hour time cache.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stocks"
                ],
                "summary": "Get Historical Stock Data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stock Symbol (e.g. RELIANCE)",
                        "name": "symbol",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.NSEHistoricalData"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.AllIndicesResponse": {
            "type": "object",
            "properties": {
                "index": {
                    "type": "string"
                },
                "indexSymbol": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "last": {
                    "type": "number"
                },
                "oneWeekAgoVal": {
                    "type": "number"
                },
                "perChange1w": {
                    "type": "number"
                },
                "perChange30d": {
                    "type": "number"
                },
                "perChange365d": {
                    "type": "number"
                },
                "percentChange": {
                    "type": "number"
                }
            }
        },
...
        "model.NSEHistoricalData": {
            "type": "object",
            "properties": {
                "chClosingPrice": {
                    "type": "number"
                },
                "chOpeningPrice": {
                    "type": "number"
                },
                "chSymbol": {
                    "type": "string"
                },
                "chTradeHighPrice": {
                    "type": "number"
                },
                "chTradeLowPrice": {
                    "type": "number"
                },
                "mtimestamp": {
                    "type": "string"
                }
            }
        },
...
        "model.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string",
                    "example": "Update successful"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "model.SectorData": {
            "type": "object",
            "properties": {
                "close": {
                    "type": "number"
                },
                "current": {
                    "type": "number"
                },
                "high": {
                    "type": "number"
                },
                "index": {
                    "type": "string"
                },
                "indexLongName": {
                    "type": "string"
                },
                "low": {
                    "type": "number"
                },
                "open": {
                    "type": "number"
                },
                "pChange": {
                    "type": "number"
                },
                "timeStamp": {
                    "type": "string"
                }
            }
        },
...
    }
}